+++ src/umpire/strategy/AlignedAllocator.cpp
@@ -30,14 +30,14 @@ AlignedAllocator::AlignedAllocator(
   }
 }
 
-void* 
+void*
 AlignedAllocator::allocate(std::size_t bytes)
 {
   std::size_t total_bytes = bytes+sizeof(void*)+m_alignment-1;
   UMPIRE_LOG(Debug, "requested: " << bytes << " actual: " << bytes+m_alignment-1);
 
   uintptr_t ptr{reinterpret_cast<uintptr_t>(m_allocator->allocate(total_bytes))};
-  uintptr_t aligned_ptr{static_cast<uintptr_t>((ptr + sizeof(void*) + (m_alignment-1)) & m_mask)}; 
+  uintptr_t aligned_ptr{static_cast<uintptr_t>((ptr + sizeof(void*) + (m_alignment-1)) & m_mask)};
   uintptr_t* header = (uintptr_t*) (aligned_ptr - sizeof(void*));
   *header = ptr;
 
@@ -45,18 +45,18 @@ AlignedAllocator::allocate(std::size_t bytes)
   return reinterpret_cast<void*>(aligned_ptr);
 }
 
-void 
+void
 AlignedAllocator::deallocate(void* ptr)
 {
   uintptr_t aligned_ptr{reinterpret_cast<uintptr_t>(ptr)};
   uintptr_t* header = (uintptr_t*) (aligned_ptr - sizeof(void*));
   void* base_ptr = reinterpret_cast<void*>(*header);
-  
+
   UMPIRE_LOG(Debug, "ptr: " << reinterpret_cast<void*>(ptr) << " base_ptr: " << reinterpret_cast<void*>(base_ptr));
   return m_allocator->deallocate(base_ptr);
 }
 
-Platform 
+Platform
 AlignedAllocator::getPlatform() noexcept
 {
   return m_allocator->getPlatform();