+++ src/umpire/strategy/AlignedAllocator.hpp
@@ -22,7 +22,7 @@ class AlignedAllocator :
         const std::string& name,
         int id,
         Allocator allocator,
-        std::size_t alignment=16); 
+        std::size_t alignment=16);
 
     void* allocate(std::size_t bytes) override;
     void deallocate(void* ptr) override;