+++ src/umpire/strategy/DynamicPoolMap.cpp
@@ -9,53 +9,35 @@
 #include "umpire/strategy/DynamicPool.hpp"
 
 #include "umpire/Allocator.hpp"
+#include "umpire/Replay.hpp"
 #include "umpire/ResourceManager.hpp"
 
+#include "umpire/strategy/mixins/AlignedAllocation.hpp"
 #include "umpire/util/Macros.hpp"
 #include "umpire/util/memory_sanitizers.hpp"
-#include "umpire/Replay.hpp"
 #include "umpire/util/backtrace.hpp"
 
 #include <cstdlib>
 #include <algorithm>
 #include <sstream>
 
-inline static std::size_t round_up(std::size_t num, std::size_t factor)
-{
-  return num + factor - 1 - (num - 1) % factor;
-}
-
 namespace umpire {
 namespace strategy {
 
-DynamicPoolMap::DynamicPoolMap(const std::string& name,
-                         int id,
-                         Allocator allocator,
-                         const std::size_t initial_alloc_bytes,
-                         const std::size_t min_alloc_bytes,
-                         const std::size_t align_bytes,
-                         CoalesceHeuristic coalesce_heuristic) noexcept :
-  AllocationStrategy(name, id),
-  m_allocator{allocator.getAllocationStrategy()},
-  m_initial_alloc_bytes{round_up(initial_alloc_bytes, align_bytes)},
-  m_min_alloc_bytes{round_up(min_alloc_bytes, align_bytes)},
-  m_align_bytes{align_bytes},
-  m_coalesce_heuristic{coalesce_heuristic},
-  m_used_map{},
-  m_free_map{},
-  m_actual_bytes{round_up(initial_alloc_bytes, align_bytes)}
+DynamicPoolMap::DynamicPoolMap(
+    const std::string& name,
+    int id,
+    Allocator allocator,
+    const std::size_t first_minimum_pool_allocation_size,
+    const std::size_t next_minimum_pool_allocation_size,
+    const std::size_t alignment,
+    CoalesceHeuristic should_coalesce) noexcept :
+  AllocationStrategy{name, id},
+  mixins::AlignedAllocation{ alignment, allocator.getAllocationStrategy() },
+  m_should_coalesce{ should_coalesce },
+  m_first_minimum_pool_allocation_size{ aligned_round_up(first_minimum_pool_allocation_size) },
+  m_next_minimum_pool_allocation_size{ aligned_round_up(next_minimum_pool_allocation_size) }
 {
-  const std::size_t bytes{round_up(initial_alloc_bytes, align_bytes)};
-#if defined(UMPIRE_ENABLE_BACKTRACE)
-  {
-    umpire::util::backtrace bt{};
-    umpire::util::backtracer<>::get_backtrace(bt);
-    UMPIRE_LOG(Info, "actual_size: " << bytes << " (prev: 0) " << umpire::util::backtracer<>::print(bt));
-  }
-#endif
-  void* ptr = m_allocator->allocate(bytes);
-  insertFree(ptr, bytes, true, bytes);
-  UMPIRE_POISON_MEMORY_REGION(m_allocator, ptr, bytes);
 }
 
 DynamicPoolMap::~DynamicPoolMap()
@@ -71,7 +53,9 @@ DynamicPoolMap::~DynamicPoolMap()
     std::size_t whole_bytes;
     std::tie(addr, is_head, whole_bytes) = rec.second;
     // Deallocate if this is a whole block
-    if (is_head && bytes == whole_bytes) deallocateBlock(addr, bytes);
+    if (is_head && bytes == whole_bytes) {
+      deallocateBlock(addr, bytes);
+    }
   }
 
   if (m_used_map.size() == 0) {
@@ -111,20 +95,20 @@ DynamicPoolMap::findFreeBlock(std::size_t bytes) const
   return iter;
 }
 
-void* DynamicPool::allocateBlock(std::size_t bytes)
+void* DynamicPoolMap::allocateBlock(std::size_t bytes)
 {
   void* ptr{nullptr};
   try {
 #if defined(UMPIRE_ENABLE_BACKTRACE)
     {
-      umpire::util::backtrace bt{};
+      umpire::util::backtrace bt;
       umpire::util::backtracer<>::get_backtrace(bt);
       UMPIRE_LOG(Info, "actual_size: " << (m_actual_bytes+bytes)
         << " (prev: " << m_actual_bytes << ") "
         << umpire::util::backtracer<>::print(bt));
     }
 #endif
-    ptr = m_allocator->allocate(bytes);
+    ptr = aligned_allocate(bytes);
   } catch (...) {
     UMPIRE_LOG(Error,
                "\n\tMemory exhausted at allocation resource. "
@@ -141,7 +125,7 @@ void* DynamicPool::allocateBlock(std::size_t bytes)
                << getInUseBlocks() << " Used Blocks\n"
       );
     try {
-      ptr = m_allocator->allocate(bytes);
+      ptr = aligned_allocate(bytes);
       UMPIRE_LOG(Error,
                  "\n\tMemory successfully recovered at resource.  Allocation succeeded\n"
         );
@@ -159,27 +143,27 @@ void* DynamicPool::allocateBlock(std::size_t bytes)
 
   UMPIRE_POISON_MEMORY_REGION(m_allocator, ptr, bytes);
 
-  // Add to count
   m_actual_bytes += bytes;
 
   return ptr;
 }
 
-void DynamicPool::deallocateBlock(void* ptr, std::size_t bytes)
+void DynamicPoolMap::deallocateBlock(void* ptr, std::size_t size)
 {
-  m_actual_bytes -= bytes;
-  m_allocator->deallocate(ptr);
+  UMPIRE_POISON_MEMORY_REGION(m_allocator, ptr, size);
+  m_actual_bytes -= size;
+  aligned_deallocate(ptr);
 }
 
-void* DynamicPool::allocate(std::size_t bytes)
+void* DynamicPoolMap::allocate(std::size_t bytes)
 {
+  bytes = aligned_round_up(bytes);
   UMPIRE_LOG(Debug, "(bytes=" << bytes << ")");
 
-  const std::size_t rounded_bytes = round_up(bytes, m_align_bytes);
   Pointer ptr{nullptr};
 
   // Check if the previous block is a match
-  const SizeMap::const_iterator iter{findFreeBlock(rounded_bytes)};
+  const SizeMap::const_iterator iter{findFreeBlock(bytes)};
 
   if (iter != m_free_map.end()) {
     // Found this acceptable address pair
@@ -188,38 +172,39 @@ void* DynamicPool::allocate(std::size_t bytes)
     std::tie(ptr, is_head, whole_bytes) = iter->second;
 
     // Add used map
-    insertUsed(ptr, rounded_bytes, is_head, whole_bytes);
+    insertUsed(ptr, bytes, is_head, whole_bytes);
 
     // Remove the entry from the free map
     const std::size_t free_size{iter->first};
     m_free_map.erase(iter);
 
-    const std::size_t left_bytes{free_size - rounded_bytes};
+    const std::size_t left_bytes{free_size - bytes};
 
     if (left_bytes > 0) {
-      insertFree(static_cast<unsigned char*>(ptr) + rounded_bytes, left_bytes,
+      insertFree(static_cast<unsigned char*>(ptr) + bytes, left_bytes,
                  false, whole_bytes);
     }
   } else {
-    // Allocate new block -- note this does not check whether alignment is met
     const std::size_t min_block_size =
-      ( m_actual_bytes == 0 ) ? m_initial_alloc_bytes : m_min_alloc_bytes;
+      ( m_actual_bytes == 0 ) ? m_first_minimum_pool_allocation_size
+                              : m_next_minimum_pool_allocation_size;
 
-    const std::size_t alloc_bytes{std::max(rounded_bytes, min_block_size)};
+    const std::size_t alloc_bytes{std::max(bytes, min_block_size)};
     ptr = allocateBlock(alloc_bytes);
 
-    // Add used
-    insertUsed(ptr, rounded_bytes, true, alloc_bytes);
+    UMPIRE_ASSERT("bytes too large" && bytes <= alloc_bytes);
 
-    const std::size_t left_bytes{alloc_bytes - rounded_bytes};
+    insertUsed(ptr, bytes, true, alloc_bytes);
+
+    const std::size_t left_bytes{alloc_bytes - bytes};
 
     // Add free
     if (left_bytes > 0)
-      insertFree(static_cast<unsigned char*>(ptr) + rounded_bytes, left_bytes,
+      insertFree(static_cast<unsigned char*>(ptr) + bytes, left_bytes,
                  false, alloc_bytes);
   }
 
-  UMPIRE_UNPOISON_MEMORY_REGION(m_allocator, ptr, rounded_bytes);
+  UMPIRE_UNPOISON_MEMORY_REGION(m_allocator, ptr, bytes);
   return ptr;
 }
 
@@ -249,7 +234,7 @@ void DynamicPoolMap::deallocate(void* ptr)
     UMPIRE_ERROR("Cound not found ptr = " << ptr);
   }
 
-  if (m_coalesce_heuristic(*this)) {
+  if (m_should_coalesce(*this)) {
     UMPIRE_LOG(Debug, this
                << " heuristic function returned true, calling coalesce()");
     do_coalesce();
@@ -410,7 +395,7 @@ std::size_t DynamicPoolMap::releaseFreeBlocks()
 
 #if defined(UMPIRE_ENABLE_BACKTRACE)
   if (released_bytes > 0) {
-    umpire::util::backtrace bt{};
+    umpire::util::backtrace bt;
     umpire::util::backtracer<>::get_backtrace(bt);
     UMPIRE_LOG(Info, "actual_size: " << m_actual_bytes
       << " (prev: " << (m_actual_bytes+released_bytes)
@@ -455,10 +440,38 @@ void DynamicPoolMap::do_coalesce()
 
     // If this removed anything from the map, re-allocate a single large chunk and insert to free map
     if (released_bytes > 0) {
-      const Pointer ptr{allocateBlock(released_bytes)};
+      Pointer ptr{allocateBlock(released_bytes)};
       insertFree(ptr, released_bytes, true, released_bytes);
     }
   }
 }
+
+DynamicPoolMap::CoalesceHeuristic
+DynamicPoolMap::percent_releasable(int percentage)
+{
+  if ( percentage < 0 || percentage > 100 ) {
+    UMPIRE_ERROR("Invalid percentage of " << percentage
+        << ", percentage must be an integer between 0 and 100");
+  }
+
+  if ( percentage == 0 ) {
+    return [=] (const DynamicPoolMap& UMPIRE_UNUSED_ARG(pool)) {
+        return false;
+    };
+  } else if ( percentage == 100 ) {
+    return [=] (const strategy::DynamicPoolMap& pool) {
+        return ( pool.getActualSize() == pool.getReleasableSize() );
+    };
+  } else {
+    float f = (float)((float)percentage / (float)100.0);
+
+    return [=] (const strategy::DynamicPoolMap& pool) {
+      // Calculate threshold in bytes from the percentage
+      const std::size_t threshold = static_cast<std::size_t>(f * pool.getActualSize());
+      return (pool.getReleasableSize() >= threshold);
+    };
+  }
+}
+
 } // end of namespace strategy
 } // end of namespace umpire