+++ src/umpire/strategy/DynamicPoolMap.hpp
@@ -8,10 +8,11 @@
 #define UMPIRE_DynamicPoolMap_HPP
 
 #include "umpire/strategy/AllocationStrategy.hpp"
-#include "umpire/strategy/DynamicPoolHeuristic.hpp"
 
+#include "umpire/strategy/mixins/AlignedAllocation.hpp"
 #include "umpire/util/MemoryMap.hpp"
 
+#include <functional>
 #include <map>
 #include <tuple>
 
@@ -32,41 +33,35 @@ namespace strategy {
  * and the minimum size controls the lower bound on all future chunk
  * allocations.
  */
-class DynamicPoolMap : public AllocationStrategy
+class DynamicPoolMap : public AllocationStrategy, private mixins::AlignedAllocation
 {
   public:
     using Pointer = void*;
 
-    /*!
-     * \brief Callback heuristic to trigger coalesce of free blocks in pool.
-     *
-     * The registered heuristic callback function will be called immediately
-     * after a deallocation() has completed from the pool.
-     */
     using CoalesceHeuristic = std::function<bool (const strategy::DynamicPoolMap&)>;
+
+    static CoalesceHeuristic percent_releasable(int percentage);
 
     /*!
      * \brief Construct a new DynamicPoolMap.
      *
      * \param name Name of this instance of the DynamicPoolMap
      * \param id Unique identifier for this instance
-     * \param initial_alloc_bytes Size the pool initially allocates
-     * \param min_alloc_bytes The minimum size of all future allocations
-     * \param coalesce_heuristic Heuristic callback function
-     * \param align_bytes Number of bytes with which to align allocation sizes
+     * \param allocator Allocation resource that pool uses
+     * \param first_minimum_pool_allocation_size Size the pool initially allocates
+     * \param next_minimum_pool_allocation_size The minimum size of all future allocations
+     * \param align_bytes Number of bytes with which to align allocation sizes (power-of-2)
+     * \param should_coalesce Heuristic for when to perform coalesce operation
      */
     DynamicPoolMap(
         const std::string& name,
         int id,
         Allocator allocator,
-        const std::size_t initial_alloc_size = (512 * 1024 * 1024),
+        const std::size_t first_minimum_pool_allocation_size = (512 * 1024 * 1024),
         const std::size_t min_alloc_size = (1 * 1024 * 1024),
         const std::size_t align_bytes = 16,
-        CoalesceHeuristic coalesce_heuristic = heuristic_percent_releasable(100)) noexcept;
+        CoalesceHeuristic should_coalesce = percent_releasable(100)) noexcept;
 
-    /*!
-     * \brief Destructs the DynamicPoolMap.
-     */
     ~DynamicPoolMap();
 
     DynamicPoolMap(const DynamicPoolMap&) = delete;
@@ -140,7 +135,7 @@ class DynamicPoolMap : public AllocationStrategy
     /*!
      * \brief Deallocate from m_allocator.
      */
-    void deallocateBlock(void* ptr, std::size_t bytes);
+    void deallocateBlock(void* ptr, std::size_t size);
 
     /*!
      * \brief Insert a block to the used map.
@@ -175,14 +170,15 @@ class DynamicPoolMap : public AllocationStrategy
 
     void do_coalesce();
 
-    strategy::AllocationStrategy* m_allocator;
-    const std::size_t m_initial_alloc_bytes;
-    const std::size_t m_min_alloc_bytes;
-    const std::size_t m_align_bytes;
-    CoalesceHeuristic m_coalesce_heuristic;
-    AddressMap m_used_map;
-    SizeMap m_free_map;
-    std::size_t m_actual_bytes;
+    AddressMap m_used_map{};
+    SizeMap m_free_map{};
+
+    CoalesceHeuristic m_should_coalesce;
+
+    const std::size_t m_first_minimum_pool_allocation_size;
+    const std::size_t m_next_minimum_pool_allocation_size;
+
+    std::size_t m_actual_bytes{0};
 };
 
 } // end of namespace strategy